aspect ratio sep test
